6.1

+++ SQLTest/src/com/gui/bean/Employee.java
@@ -5,13 +5,13 @@ import java.sql.Date;
 public class Employee {
 	private String last_name;
 	private String phone_number;
-	private int salary;
+	private double salary;
 	private Date hiredate;
 	public Employee() {
 		super();
 		// TODO 自动生成的构造函数存根
 	}
-	public Employee(String last_name, String phone_number, int salary, Date hiredate) {
+	public Employee(String last_name, String phone_number, double salary, Date hiredate) {
 		super();
 		this.last_name = last_name;
 		this.phone_number = phone_number;
@@ -30,7 +30,7 @@ public class Employee {
 	public void setPhone_number(String phone_number) {
 		this.phone_number = phone_number;
 	}
-	public int getSalary() {
+	public double getSalary() {
 		return salary;
 	}
 	public void setSalary(int salary) {
